test(*): fix server side test for modified submitReview method and add new tests
