+++ packages/server/api/__tests__/question.api.test.js
@@ -18,21 +18,7 @@ const {
   createIdentity,
 } = require('../../models/__tests__/__helpers__/users')
 
-const {
-  createQuestion,
-  updateReviewerPool,
-} = require('../../controllers/question.controllers')
-
-const {
-  createEmptyQuestion,
-} = require('../../controllers/__tests__/__helpers__/questions')
-
-const {
-  acceptOrRejectInvitation,
-} = require('../../controllers/team.controllers')
-
-const { REVIEWER_STATUSES } = require('../../controllers/constants')
-const { submitReview } = require('../../controllers/review.controller')
+const { createQuestion } = require('../../controllers/question.controllers')
 
 // gql queries
 const GET_AUTHOR_DASHBOARD = `
@@ -1562,129 +1548,5 @@ describe('Question API authorization', () => {
     expect(isAdmin).toBe(true)
     expect(result.errors).toBe(undefined)
     expect(result.data).not.toBe(null)
-  })
-
-  it('fetches the correct reviewer status', async () => {
-    const user = await createUser()
-    const question = await createEmptyQuestion()
-    const editor = await createUser()
-    const handlingEditor1 = await createUser()
-    const handlingEditor2 = await createUser()
-
-    await createIdentity(editor, internet.email(), false, null)
-    await createIdentity(handlingEditor1, internet.email(), false, null)
-    await createIdentity(handlingEditor2, internet.email(), false, null)
-
-    let questionVersion = await QuestionVersion.findOne({
-      questionId: question.id,
-    })
-
-    const editorTeam = await Team.insert({
-      role: 'editor',
-      global: true,
-      displayName: 'Managing Editor',
-    })
-
-    await Team.updateMembershipByTeamId(editorTeam.id, [editor.id])
-
-    const handlingEditorTeam = await Team.insert({
-      role: 'handlingEditor',
-      displayName: 'Handling Editor',
-      objectId: questionVersion.questionId,
-      objectType: 'question',
-    })
-
-    await Team.updateMembershipByTeamId(handlingEditorTeam.id, [
-      handlingEditor1.id,
-      handlingEditor2.id,
-    ])
-
-    questionVersion = await QuestionVersion.findOne({
-      questionId: question.id,
-    })
-
-    const reviewer = await createUser()
-    await createIdentity(reviewer, internet.email(), false, null)
-
-    const userTestServer = await createGraphQLServer()
-
-    let response = await userTestServer.executeOperation(
-      {
-        query: GET_QUESTION,
-        variables: {
-          id: question.id,
-          published: false,
-        },
-      },
-      {
-        contextValue: {
-          userId: user.id,
-        },
-      },
-    )
-
-    let result = response.body.singleResult
-
-    const [userResultVersion] = result.data.question.versions
-
-    expect(userResultVersion.reviewerStatus).toBe(null)
-    expect(userResultVersion.reviews).toHaveLength(0)
-
-    await updateReviewerPool(questionVersion.id, [reviewer.id])
-    await acceptOrRejectInvitation(questionVersion.id, true, null, reviewer.id)
-
-    const reviewerTestServer = await createGraphQLServer()
-
-    response = await reviewerTestServer.executeOperation(
-      {
-        query: GET_QUESTION,
-        variables: {
-          id: question.id,
-          published: false,
-        },
-      },
-      {
-        contextValue: {
-          userId: reviewer.id,
-        },
-      },
-    )
-
-    result = response.body.singleResult
-    const [reviewerResultVersion] = result.data.question.versions
-
-    expect(reviewerResultVersion.reviewerStatus).toBe(
-      REVIEWER_STATUSES.accepted,
-    )
-    expect(reviewerResultVersion.reviews).toHaveLength(1)
-    expect(reviewerResultVersion.reviews[0].content).toBeFalsy()
-
-    const reviewContent = 'Looks good I guess...'
-
-    await submitReview(questionVersion.id, reviewContent, reviewer.id)
-
-    response = await reviewerTestServer.executeOperation(
-      {
-        query: GET_QUESTION,
-        variables: {
-          id: question.id,
-          published: false,
-        },
-      },
-      {
-        contextValue: {
-          userId: reviewer.id,
-        },
-      },
-    )
-
-    result = response.body.singleResult
-    const [reviewerResultVersion2] = result.data.question.versions
-
-    expect(reviewerResultVersion2.reviewerStatus).toBe(
-      REVIEWER_STATUSES.accepted,
-    )
-    expect(reviewerResultVersion2.reviews).toHaveLength(1)
-    expect(reviewerResultVersion2.reviews[0].content).toBe(reviewContent)
   })
 })
