+++ packages/server/controllers/__tests__/review.controller.test.js
@@ -13,8 +13,28 @@ const {
   changeAmountOfReviewers,
 } = require('../question.controllers')
 
-const { inviteMaxReviewers, submitReview } = require('../review.controller')
+const {
+  inviteMaxReviewers,
+  submitReview,
+  saveReview,
+} = require('../review.controller')
+
 const { createEmptyQuestion } = require('./__helpers__/questions')
+
+const reviewResponses = {
+  barriers: 'A',
+  concerns: ['grammatical'],
+  hasIssues: false,
+  difficulty: 'appropriate',
+  suggestions: 'qqqq',
+  issuesDetails: 'ISSUES',
+  otherConcerns: 'O',
+  clarityConcerns: 'C',
+  answeredCorrectly: true,
+  feedbackEvaluation: 'hasIssues',
+  grammaticalConcerns: 'Grammatical',
+  feedbackIssuesDetails: 'qq',
+}
 
 describe('Review Controller', () => {
   beforeEach(clearDb)
@@ -53,7 +73,7 @@ describe('Review Controller', () => {
     invitedMembers.forEach(member => expect(ids).toContain(member.userId))
   })
 
-  test('submitReview creates a review for the given user', async () => {
+  test('saveReview updates responses', async () => {
     const question = await createEmptyQuestion()
     const editor = await createUser()
     const handlingEditor1 = await createUser()
@@ -93,26 +113,95 @@ describe('Review Controller', () => {
 
     await createIdentity(reviewer, internet.email(), false, null)
 
-    let review = await Review.findOne({
+    await Review.insert({
+      questionVersionId: questionVersion.id,
+      reviewerId: reviewer.id,
+      status: {
+        pending: true,
+        submitted: false,
+      },
+    })
+
+    const review = await Review.findOne({
       questionVersionId: questionVersion.id,
       reviewerId: reviewer.id,
     })
 
-    expect(review).toBeFalsy()
+    expect(review.responses).toMatchObject({})
 
-    const reviewContent =
-      'Yeah, looks alright.\nMaybe change some formatting stuff.'
-
-    const reviewId = await submitReview(
+    const updatedReview = await saveReview(
       questionVersion.id,
-      reviewContent,
       reviewer.id,
+      JSON.stringify(reviewResponses),
     )
 
-    expect(reviewId).not.toBeFalsy()
-    review = await Review.findById(reviewId)
+    expect(updatedReview.responses).toMatchObject(reviewResponses)
+  })
 
-    expect(review.id).toBe(reviewId)
-    expect(review.content).toBe(reviewContent)
+  test('submitReview updates review status for user/questionVersion combination', async () => {
+    const question = await createEmptyQuestion()
+    const editor = await createUser()
+    const handlingEditor1 = await createUser()
+    const handlingEditor2 = await createUser()
+
+    await createIdentity(editor, internet.email(), false, null)
+
+    await createIdentity(handlingEditor1, internet.email(), false, null)
+
+    await createIdentity(handlingEditor2, internet.email(), false, null)
+
+    const questionVersion = await QuestionVersion.findOne({
+      questionId: question.id,
+    })
+
+    const editorTeam = await Team.insert({
+      role: 'editor',
+      global: true,
+      displayName: 'Managing Editor',
+    })
+
+    await Team.updateMembershipByTeamId(editorTeam.id, [editor.id])
+
+    const handlingEditorTeam = await Team.insert({
+      role: 'handlingEditor',
+      displayName: 'Handling Editor',
+      objectId: questionVersion.questionId,
+      objectType: 'question',
+    })
+
+    await Team.updateMembershipByTeamId(handlingEditorTeam.id, [
+      handlingEditor1.id,
+      handlingEditor2.id,
+    ])
+
+    const reviewer = await createUser()
+
+    await createIdentity(reviewer, internet.email(), false, null)
+
+    await Review.insert({
+      questionVersionId: questionVersion.id,
+      reviewerId: reviewer.id,
+      status: {
+        pending: true,
+        submitted: false,
+      },
+      // insert responses as well
+      responses: reviewResponses,
+    })
+
+    const review = await Review.findOne({
+      questionVersionId: questionVersion.id,
+      reviewerId: reviewer.id,
+    })
+
+    expect(review.status).toMatchObject({ pending: true, submitted: false })
+
+    const reviewId = await submitReview(questionVersion.id, reviewer.id)
+    const submittedReview = await Review.findById(reviewId)
+
+    expect(submittedReview.status).toMatchObject({
+      pending: false,
+      submitted: true,
+    })
   })
 })
